requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.9.0            # via tensorflow-datasets
attrs==19.3.0             # via tensorflow-datasets
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via networkx
dill==0.3.1.1             # via tensorflow-datasets
future==0.18.2            # via tensorflow-datasets
googleapis-common-protos==1.51.0  # via tensorflow-metadata
h5py==2.10.0              # via keras-applications
idna==2.9                 # via requests
image-classifiers==1.0.0  # via -r requirements.in
imageio==2.8.0            # via imgaug, scikit-image
imgaug==0.4.0             # via -r requirements.in
keras-applications==1.0.8  # via image-classifiers
kiwisolver==1.2.0         # via matplotlib
matplotlib==3.2.1         # via -r requirements.in, imgaug, scikit-image
networkx==2.4             # via scikit-image
numpy==1.18.3             # via h5py, imageio, imgaug, keras-applications, matplotlib, opencv-python, pywavelets, scipy, tensorflow-datasets
opencv-python==4.2.0.34   # via imgaug
pillow==7.1.2             # via imageio, imgaug, scikit-image
progressbar2==3.51.0      # via -r requirements.in
promise==2.3              # via tensorflow-datasets
protobuf==3.11.3          # via googleapis-common-protos, tensorflow-datasets, tensorflow-metadata
pyparsing==2.4.7          # via matplotlib
python-dateutil==2.8.1    # via matplotlib
python-utils==2.4.0       # via progressbar2
pywavelets==1.1.1         # via scikit-image
requests==2.23.0          # via tensorflow-datasets
scikit-image==0.16.2      # via imgaug
scipy==1.4.1              # via imgaug, scikit-image
shapely==1.7.0            # via imgaug
six==1.14.0               # via absl-py, cycler, h5py, imgaug, progressbar2, promise, protobuf, python-dateutil, python-utils, tensorflow-datasets
tensorflow-datasets==3.0.0  # via -r requirements.in
tensorflow-metadata==0.21.2  # via tensorflow-datasets
termcolor==1.1.0          # via tensorflow-datasets
tqdm==4.45.0              # via tensorflow-datasets
urllib3==1.25.9           # via requests
wrapt==1.12.1             # via tensorflow-datasets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
